requirements.txt
bcrypt==4.3.0
cffi==2.0.0
cfgv==3.4.0
cryptography==46.0.1
distlib==0.4.0
dotenv==0.9.9
et_xmlfile==2.0.0
filelock==3.19.1
identify==2.6.14
invoke==2.2.0
librt==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
mypy==1.19.0
mypy_extensions==1.1.0
netmiko==4.6.0
nodeenv==1.9.1
ntc_templates==8.0.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
pandas-stubs==2.3.3.251201
paramiko==4.0.0
pathspec==0.12.1
platformdirs==4.4.0
pre_commit==4.3.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.13.1
scp==0.15.0
six==1.17.0
textfsm==2.1.0
types-pytz==2025.2.0.20251108
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
virtualenv==20.34.0
